add validasi fungsi hapus rpjpd

+++ public/class-wp-eval-sakip-public.php
@@ -20158,128 +20158,6 @@ class Wp_Eval_Sakip_Public extends Wp_Eval_Sakip_Verify_Dokumen
 		die(json_encode($ret));
 	}
 
-	public function esakip_singkron_rpjpd_sipd_lokal()
-	{
-		global $wpdb;
-		$ret = array(
-			'status'    => 'success',
-			'message'   => 'Berhasil mengambil data RPD dari data SIPD lokal!'
-		);
-		if (!empty($_POST)) {
-			if (!empty($_POST['api_key']) && $_POST['api_key'] == get_option(ESAKIP_APIKEY)) {
-				$sql = "
-                    select 
-                        * 
-                    from data_rpd_tujuan
-                    where active=1
-                ";
-				$tujuan_all = $wpdb->get_results($sql, ARRAY_A);
-				foreach ($tujuan_all as $tujuan) {
-					$table = 'esakip_rpjpd_visi';
-					$id_cek_visi = $wpdb->get_var("
-                        SELECT id from $table 
-                        where visi_teks='{$tujuan['visi_teks']}'
-                    ");
-					$data = array(
-						'visi_teks' => $tujuan['visi_teks'],
-						'update_at' => $tujuan['update_at']
-					);
-					if (!empty($id_cek_visi)) {
-						$wpdb->update($table, $data, array('id' => $id_cek_visi));
-					} else {
-						$wpdb->insert($table, $data);
-						$id_cek_visi = $wpdb->insert_id;
-					}
-
-					$table = 'esakip_rpjpd_misi';
-					$id_cek_misi = $wpdb->get_var("
-                        SELECT id from $table 
-                        where misi_teks='{$tujuan['misi_teks']}'
-                            AND id_visi='$id_cek_visi'
-                    ");
-					$data = array(
-						'id_visi' => $id_cek_visi,
-						'misi_teks' => $tujuan['misi_teks'],
-						'urut_misi' => $tujuan['urut_misi'],
-						'update_at' => $tujuan['update_at']
-					);
-					if (!empty($id_cek_misi)) {
-						$wpdb->update($table, $data, array('id' => $id_cek_misi));
-					} else {
-						$wpdb->insert($table, $data);
-						$id_cek_misi = $wpdb->insert_id;
-					}
-
-					$table = 'esakip_rpjpd_sasaran';
-					$id_cek_sasaran = $wpdb->get_var("
-                        SELECT id from $table 
-                        where saspok_teks='{$tujuan['saspok_teks']}'
-                            AND id_misi='$id_cek_misi'
-                    ");
-					$data = array(
-						'id_misi' => $id_cek_misi,
-						'saspok_teks' => $tujuan['saspok_teks'],
-						'urut_saspok' => $tujuan['urut_saspok'],
-						'update_at' => $tujuan['update_at']
-					);
-					if (!empty($id_cek_sasaran)) {
-						$wpdb->update($table, $data, array('id' => $id_cek_sasaran));
-					} else {
-						$wpdb->insert($table, $data);
-						$id_cek_sasaran = $wpdb->insert_id;
-					}
-
-					$table = 'esakip_rpjpd_kebijakan';
-					$id_cek_kebijakan = $wpdb->get_var("
-                        SELECT id from $table 
-                        where kebijakan_teks='{$tujuan['kebijakan_teks']}'
-                            AND id_saspok='$id_cek_sasaran'
-                    ");
-					$data = array(
-						'id_saspok' => $id_cek_sasaran,
-						'kebijakan_teks' => $tujuan['kebijakan_teks'],
-						'update_at' => $tujuan['update_at']
-					);
-					if (!empty($id_cek_kebijakan)) {
-						$wpdb->update($table, $data, array('id' => $id_cek_kebijakan));
-					} else {
-						$wpdb->insert($table, $data);
-						$id_cek_kebijakan = $wpdb->insert_id;
-					}
-
-					$table = 'esakip_rpjpd_isu';
-					$id_cek_isu = $wpdb->get_var("
-                        SELECT id from $table 
-                        where isu_teks='{$tujuan['isu_teks']}'
-                            AND id_kebijakan='$id_cek_kebijakan'
-                    ");
-					$data = array(
-						'id_kebijakan' => $id_cek_kebijakan,
-						'isu_teks' => $tujuan['isu_teks'],
-						'update_at' => $tujuan['update_at']
-					);
-					if (!empty($id_cek_isu)) {
-						$wpdb->update($table, $data, array('id' => $id_cek_isu));
-					} else {
-						$wpdb->insert($table, $data);
-						$id_cek_isu = $wpdb->insert_id;
-					}
-				}
-			} else {
-				$ret = array(
-					'status' => 'error',
-					'message'   => 'Api Key tidak sesuai!'
-				);
-			}
-		} else {
-			$ret = array(
-				'status' => 'error',
-				'message'   => 'Format tidak sesuai!'
-			);
-		}
-		die(json_encode($ret));
-	}
-
 	public function esakip_get_rpjpd($res = false, $id_jadwal_history = false)
 	{
 		global $wpdb;
@@ -20646,18 +20524,71 @@ class Wp_Eval_Sakip_Public extends Wp_Eval_Sakip_Verify_Dokumen
 		if (!empty($_POST)) {
 			if (!empty($_POST['api_key']) && $_POST['api_key'] == get_option(ESAKIP_APIKEY)) {
 				$table = '';
+				$cek_id = '';
 				if ($_POST['table'] == 'esakip_rpjpd_visi') {
-					$table = $_POST['table'];
-					$wpdb->delete($table, array('id' => $_POST['id']));
+					$cek_id = $wpdb->get_var(
+						$wpdb->prepare("
+							SELECT id_visi
+							FROM esakip_rpjpd_misi
+							WHERE id_visi = %d
+						", $_POST['id'])
+					);
+					if (empty($cek_id)) {
+						$table = $_POST['table'];
+						$wpdb->delete($table, array('id' => $_POST['id']));
+					} else {
+						$ret['status'] = 'error';
+						$ret['message'] = 'Visi memiliki Misi aktif! Mohon hapus Misi turunannya terlebih dahulu!';
+						die(json_encode($ret));
+					}
 				} else if ($_POST['table'] == 'esakip_rpjpd_misi') {
-					$table = $_POST['table'];
-					$wpdb->delete($table, array('id' => $_POST['id']));
+					$cek_id = $wpdb->get_var(
+						$wpdb->prepare("
+							SELECT id_misi
+							FROM esakip_rpjpd_sasaran
+							WHERE id_misi = %d
+						", $_POST['id'])
+					);
+					if (empty($cek_id)) {
+						$table = $_POST['table'];
+						$wpdb->delete($table, array('id' => $_POST['id']));
+					} else {
+						$ret['status'] = 'error';
+						$ret['message'] = 'Misi memiliki Sasaran aktif! Mohon hapus Sasaran turunannya terlebih dahulu!';
+						die(json_encode($ret));
+					}
 				} else if ($_POST['table'] == 'esakip_rpjpd_sasaran') {
-					$table = $_POST['table'];
-					$wpdb->delete($table, array('id' => $_POST['id']));
+					$cek_id = $wpdb->get_var(
+						$wpdb->prepare("
+							SELECT id_saspok
+							FROM esakip_rpjpd_kebijakan
+							WHERE id_saspok = %d
+						", $_POST['id'])
+					);
+					if (empty($cek_id)) {
+						$table = $_POST['table'];
+						$wpdb->delete($table, array('id' => $_POST['id']));
+					} else {
+						$ret['status'] = 'error';
+						$ret['message'] = 'Sasaran memiliki Kebijakan aktif! Mohon hapus Kebijakan turunannya terlebih dahulu!';
+						die(json_encode($ret));
+					}
 				} else if ($_POST['table'] == 'esakip_rpjpd_kebijakan') {
-					$table = $_POST['table'];
-					$wpdb->delete($table, array('id' => $_POST['id']));
+					$cek_id = $wpdb->get_var(
+						$wpdb->prepare("
+							SELECT id_kebijakan
+							FROM esakip_rpjpd_isu
+							WHERE id_kebijakan = %d
+						", $_POST['id'])
+					);
+					if (empty($cek_id)) {
+						$table = $_POST['table'];
+						$wpdb->delete($table, array('id' => $_POST['id']));
+					} else {
+						$ret['status'] = 'error';
+						$ret['message'] = 'Kebijakan memiliki Isu aktif! Mohon hapus Isu turunannya terlebih dahulu!';
+						die(json_encode($ret));
+					}
 				} else if ($_POST['table'] == 'esakip_rpjpd_isu') {
 					$table = $_POST['table'];
 					$wpdb->delete($table, array('id' => $_POST['id']));
